Provide a valid python2.exe in the path
Copy over C:\Python27\python.exe to C:\Python27\python2.exe so
QtWebengine can have a working python2.exe. chromium requires the python
executable to be called python2.exe, and it discovers it from the PATH.

Without making this copy, it finds the python2.exe that's located in the
Git installation at C:\Program Files\Git\mingw64\bin - and that
QtWebEngine barfs on because it can't handle spaces in the path, despite
having some logic to quote the path -- it just turns out that it quotes
it so it can work with system() calls rather than with a shell.