Cherry-pick patch-set2 of https://codereview.qt-project.org/c/qt/qtbase/+/267755
Pick up a fix that has landed in the qtbase 5.12 branch.
It has not yet landed in qt5 base repo, so this cherry-pick is
necessary.

I opted to not just include all new commits on qt 5.12 branch of qtbase,
and instead only selectively pick this one commit that is needed by
https://jira.autodesk.com/browse/MAYA-99420. It may also affect a few
other Maya bugs.

Affects: MAYA-99420