Compilation fixes for QT_NO_SSL and MAYA-114941, -114065 (bad sha)
Fixes bad sha1 of qtbase branch for below changes.

Fix Qt builds that specify not to build in SSL support.
This was broken in Qt 5.15.2, and required a cherry pick from dev to
fix.

Fix for Compile-time syntax error introduced in cherry-pick for
MAYA-114065/QTBUG-98856.

C++17->C++14 support backport for tst_qfontcache fixing build with
tests.