Find Qt Version # when QTVERSION is not a version, add error handling
This looks at the qt5 codebase to determine what the Qt version number
is when the QTVERSION environment variable passed in is not a version
number (occurs when the branch name doesn't match the expected format).

It will still be used to determine the Qt installation prefix, but not
for anything else in the script.

Add bash script safety / error handling:
    * Terminate on dereferencing undefined environment variables.
    * Terminate on commands that return nonzero return codes.

Add robustness in the name of the source directory. No longer does the
base of the repository need to be in a directory called 'src'. Instead,
it is determined based on directory in which the the script is located.